v4.0.2: More updates: save solution to file, memory optimization
 - change many int to short type
 - used user defined bool type
 - made folder to save solutions to

+++ SudokuSolver/algorithms.h
@@ -2,14 +2,14 @@
 
 #include "board_globals.h"
 
-int naked_single_search(int board[LEN][LEN], int candidates[LEN][LEN][LEN]);
+int naked_single_search(short board[LEN][LEN], short candidates[LEN][LEN][LEN]);
 
-int hidden_single_search(int board[LEN][LEN], int candidates[LEN][LEN][LEN]);
+int hidden_single_search(short board[LEN][LEN], short candidates[LEN][LEN][LEN]);
 
-int locked_candidate_search(int candidates[LEN][LEN][LEN]);
+int locked_candidate_search(short candidates[LEN][LEN][LEN]);
 
-void naked_hidden_sets_search(int candidates[LEN][LEN][LEN], int set_size, int* num_naked, int* num_hidden);
+void naked_hidden_sets_search(short candidates[LEN][LEN][LEN], int set_size, int* num_naked, int* num_hidden);
 
-int basic_fish_search(int candidates[LEN][LEN][LEN], int fish_size);
+int basic_fish_search(short candidates[LEN][LEN][LEN], int fish_size);
 
-int randomized_value_board_search(int board[LEN][LEN], int candidates[LEN][LEN][LEN]);
+int randomized_value_board_search(short board[LEN][LEN], short candidates[LEN][LEN][LEN]);
